Add Statistics feature

+++ src/main/java/com/szagoret/n26/statistics/models/Statistic.java
@@ -13,5 +13,5 @@ public class Statistic {
     private Double avg;
     private Double max;
     private Double min;
-    private Double count;
+    private Long count;
 }
